requirements.txt
streamlit>=1.44
pandas>=2.2
openpyxl>=3.1
python-dotenv>=1.0
google-api-python-client>=2.165
youtube-transcript-api>=1.1
isodate>=0.7
langchain-core>=0.3
langchain-community>=0.3
langchain-text-splitters>=0.3
faiss-cpu>=1.10
requests>=2.32
openai>=1.75
